Separate module metadata into modules.rs
